# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kuben API",
    "description": "Kuben — Kubernetes-native PaaS control plane.",
    "license": {
      "name": "Apache-2.0"
    },
    "version": "1.1.2"
  },
  "paths": {
...
    "/api/v1/projects/{project}/environments/{environment}/registries": {
      "get": {
        "tags": [
          "secrets"
        ],
        "summary": "List the registry logins of an environment (never their passwords).",
        "operationId": "listRegistryLogins",
        "parameters": [
          {
            "name": "project",
            "in": "path",
            "description": "Project name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "environment",
            "in": "path",
            "description": "Environment short name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RegistryLoginDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project}/environments/{environment}/registries/{name}": {
      "put": {
        "tags": [
          "secrets"
        ],
        "summary": "Set the login of a registry: a new revision, used by the next runs.",
        "operationId": "putRegistryLogin",
        "parameters": [
          {
            "name": "project",
            "in": "path",
            "description": "Project name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "environment",
            "in": "path",
            "description": "Environment short name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "description": "Login name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PutRegistryLogin"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistryLoginDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "The name or registry is taken",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "secrets"
        ],
        "summary": "Delete a registry login. Runs accepted with it keep their revision.",
        "operationId": "deleteRegistryLogin",
        "parameters": [
          {
            "name": "project",
            "in": "path",
            "description": "Project name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "environment",
            "in": "path",
            "description": "Environment short name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "description": "Login name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "An app reads it as a secret",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Problem": {
        "type": "object",
        "description": "Problem Details body.",
        "required": [
          "code",
          "title",
          "status"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine-readable code, e.g. `forbidden`.",
            "example": "forbidden"
          },
          "title": {
            "type": "string",
            "description": "Short human readable title."
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "HTTP status.",
            "minimum": 0
          },
          "detail": {
            "type": [
              "string",
              "null"
            ],
            "description": "Details, if safe to expose."
          }
        }
      },
...
      "PutRegistryLogin": {
        "type": "object",
        "required": [
          "registry",
          "username",
          "password"
        ],
        "properties": {
          "registry": {
            "type": "string",
            "description": "`ghcr.io`, `docker.io`, `registry.example.com:5000`."
          },
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "description": "A password or access token; write-only."
          }
        },
        "additionalProperties": false
      },
...
      "RegistryLoginDto": {
        "type": "object",
        "required": [
          "name",
          "registry",
          "revision",
          "revoked",
          "updatedAt"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "registry": {
            "type": "string",
            "description": "The registry as image references name it (`ghcr.io`, `docker.io`)."
          },
          "revision": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "revoked": {
            "type": "boolean",
            "description": "The current revision is revoked: images from the registry are not\ndeployed until a new login is set."
          },
          "updatedAt": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}